# docs/swagger.yaml
basePath: /api
definitions:
...
  ApiResponse:
    description: Generic API response
    properties:
      data: {}
      responseKey:
        type: string
    type: object
...
  LoanRequestCount:
    description: LoanRequestCount represents the count of LoanRequest.
    properties:
      count:
        description: Count is the number of LoanRequest.
        type: integer
    type: object
  LoanRequestDemande:
    description: LoanRequestDemande represents a demande information about a specific
      .
    properties:
      interestRate:
        description: The interest rate of the loan request
        type: number
      loanAmount:
        description: The amount of the loan request
        type: number
      loanDuration:
        description: The duration of the loan request
        type: string
      pathDocument:
        description: The path document for the loan request
        type: string
      reasonForLoan:
        description: Reason or explanation for requesting the loan.
        type: string
    type: object
  LoanRequestDetails:
    description: LoanRequestDetails represents detailed information about a specific
      LoanRequest.
    properties:
      createdAt:
        description: CreatedAt is the timestamp indicating when the loanRequests was
          created.
        type: string
      id:
        description: Unique identifier for the LoanRequest
        type: string
      interestRate:
        description: The interest rate of the loan request
        type: number
      loanAmount:
        description: The amount of the loan request
        type: number
      loanDuration:
        description: The duration of the loan request
        type: string
      pathDocument:
        description: The path document for the loan request
        type: string
      reasonForLoan:
        description: Reason or explanation for requesting the loan.
        type: string
      status:
        description: Status of the LoanRequest request (e.g., pending, approved, rejected).
        type: string
      userID:
        description: User ID associated with the LoanRequest
        type: string
    required:
    - status
    type: object
  LoanRequestPagination:
    description: LoanRequestPagination represents the paginated list of LoanRequest.
    properties:
      items:
        description: Items is a slice containing individual LoanRequest details.
        items:
          $ref: '#/definitions/LoanRequestDetails'
        type: array
      limit:
        description: Limit is the maximum number of items per page in the pagination.
        type: integer
      page:
        description: Page is the current page number in the pagination.
        type: integer
      totalCount:
        description: TotalCount is the total number of LoanRequest in the entire list.
        type: integer
    type: object
...
  loanRequests.LoanRequestIn:
    description: LoanRequestIn represents the input structure for updating the status
      of LoanRequest.
    properties:
      status:
        description: Status of the LoanRequest request (e.g., pending, approved, rejected).
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: h.manai@tunisiancloud.com
    name: Hedi Manai
  description: A basic template for managing your project using Golang REST APIs.
  title: Project Name
  version: "0.1"
paths:
...
  /loanRequests:
    post:
      consumes:
      - application/json
      description: Create a new loanRequests.
      parameters:
      - description: LoanRequests query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/LoanRequestDemande'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Create an loanRequests demande
      tags:
      - LoanRequests
  /loanRequests/{userID}:
    get:
      description: Get all LoanRequest for a specific user .
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/LoanRequestPagination'
            type: array
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get LoanRequests
      tags:
      - LoanRequests
  /loanRequests/user/{userID}/{ID}:
    delete:
      consumes:
      - application/json
      description: Delete one loanRequests.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: LoanRequests ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete loanRequests
      tags:
      - LoanRequests
    get:
      description: Get one loanRequests.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: LoanRequests ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LoanRequestDetails'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get loanRequests
      tags:
      - LoanRequests
    put:
      consumes:
      - application/json
      description: Update one loanRequests.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: LoanRequests ID
        in: path
        name: ID
        required: true
        type: string
      - description: LoanRequests query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/loanRequests.LoanRequestIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ApiResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Update loanRequests
      tags:
      - LoanRequests
  /loanRequests/user/{userID}/count:
    get:
      description: Get all loanRequests count.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LoanRequestCount'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ApiResponse'
      security:
      - ApiKeyAuth: []
      summary: Get loanRequests count
      tags:
      - LoanRequests
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"